requirements.txt
django
django-simpleui
django-import-export